Fix and refactor checkWinner code
There was a winner for simple patterns(when number of Xs/Os <= 3) but
when there was more, even if someone won, the game wouldn't end. That
is because the formatted winning boards contained more Xs or Os than
the combinations so checking if a formatted board was the same as a
combination was incorrect. That's why I refactored the code to compare
if a non-empty space in a combination is the same as the formatted
board's space, if not then that board is not a winner.